Trimmed down spin lock implementation for rmem.
This removes various things from the mutex code including default
(non-spin) mutexes, thread_lock, debugging.  It also uses a
simplified mtx_init which does not interface with WITNESS or the
default lock_init() but just initializes the mutex fields.

For now, this uses a hack for curthread which depends on access to
the local 'tid' variable (and thus only works for inline macros in
thread functions).  Eventually this should be changed to use the TLS
register instead and have threads initialize curthread that way
(or initialize a struct pcpu or the like).